Unify reflow training and data generation
